Refactoring calculation classes
