+++ WebApp/src/uk/ac/exeter/QuinCe/data/Files/RawDataFile.java
@@ -11,9 +11,9 @@ import java.util.Arrays;
 import java.util.Calendar;
 import java.util.List;
 
-import uk.ac.exeter.QuinCe.data.DateTimeParseException;
 import uk.ac.exeter.QuinCe.data.Instrument.Instrument;
 import uk.ac.exeter.QuinCe.data.Instrument.InstrumentException;
+import uk.ac.exeter.QuinCe.utils.DateTimeParseException;
 import uk.ac.exeter.QuinCe.utils.DateTimeUtils;
 
 /**
